feat: secure token storage, E2EE verification, and UI refactor
Security:
- Migrate access tokens from SharedPreferences to flutter_secure_storage
- Add automatic legacy token migration on app launch
- Remove plaintext tokens from session metadata

E2EE:
- Add Rust-side verification request event handler
- Track verification session state (flow_id, device_id, accepted)
- Support incoming to-device key verification requests

Sync:
- Introduce SyncTask struct with JoinHandle for clean abort
- Extract stop_sync_task() helper for account switching
- Replace tokio::spawn with std::thread::spawn for log sink

UI:
- Simplify AppAvatar: remove custom _ImageCache and dart:io HttpClient
- Use StatefulBuilder for image loading state
- Invalidate avatar on active user change

Init:
- Move bootstrap (migrateLegacySession + loadAllSessions) before runApp
- Pass hasSessions to _AppRoot via constructor, remove async _bootstrap
- Set sessionReadyProvider immediately when no sessions exist

Deps:
- Add flutter_secure_storage to pubspec.yaml

+++ lib/pages/settings/log_viewer_page.dart
@@ -265,7 +265,7 @@ class _LogEntryTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final time = DateTime.fromMillisecondsSinceEpoch(entry.timestamp);
+    final time = DateTime.fromMillisecondsSinceEpoch(entry.timestamp.toInt());
     final timeStr =
         '${time.hour.toString().padLeft(2, '0')}:${time.minute.toString().padLeft(2, '0')}:${time.second.toString().padLeft(2, '0')}.${time.millisecond.toString().padLeft(3, '0')}';
